Remove unused `SendConfirmation.tsx` and move `SendConfirmation_updated.tsx`
